proje_dosyalari/segmentation-service/requirements.txt
fastapi
uvicorn[standard]
loguru
PyYAML
pydantic
